convert_ast2cmdinvo.c 内の関数を一部別ファイルに移動
